Issue 33 feature/#51 info fragment bookmark review comment api

+++ ita/app/src/main/java/com/ita/poppop/view/empty/info/review/detail/reply/InfoReviewDetailReplyFragment.kt
@@ -3,6 +3,7 @@ package com.ita.poppop.view.empty.info.review.detail.reply
 
 import android.graphics.Rect
 import android.util.Log
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
@@ -11,21 +12,17 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.ita.poppop.R
 import com.ita.poppop.base.BaseFragment
 import com.ita.poppop.data.remote.repository.popup.CommentRepositoryImpl
-import com.ita.poppop.data.remote.repository.popup.ReviewRepositoryImpl
 import com.ita.poppop.databinding.FragmentInfoReviewDetailReplyBinding
 import com.ita.poppop.util.ViewModelFactory
 import com.ita.poppop.util.remote.RetrofitClient
 import com.ita.poppop.view.empty.info.review.comment.InfoReviewCommentDeleteBottomSheet
-import com.ita.poppop.view.empty.info.review.comment.InfoReviewCommentRVAdapter
-import com.ita.poppop.view.empty.info.review.detail.InfoReviewDeleteBottomSheet
-import com.ita.poppop.view.empty.info.review.detail.InfoReviewDetailFragmentArgs
-import com.ita.poppop.view.empty.info.review.detail.InfoReviewDetailFragmentDirections
-import com.ita.poppop.view.empty.info.review.detail.InfoReviewDetailViewHolder
-import com.ita.poppop.view.empty.info.review.detail.InfoReviewDetailViewModel
+import com.ita.poppop.view.empty.info.review.comment.InfoReviewCommentReportBottomSheet
+import com.ita.poppop.viewmodel.MainAViewModel
 
 class InfoReviewDetailReplyFragment : BaseFragment<FragmentInfoReviewDetailReplyBinding>(R.layout.fragment_info_review_detail_reply){
 
     private val infoReviewDetailReplyArgs: InfoReviewDetailReplyFragmentArgs by navArgs()
+    val mainViewModel: MainAViewModel by activityViewModels()
 
     private lateinit var infoReviewDetailReplyViewModel: InfoReviewDetailReplyViewModel
     private lateinit var infoReviewCommentDetailViewHolder: InfoReviewCommentDetailViewHolder
@@ -45,18 +42,20 @@ class InfoReviewDetailReplyFragment : BaseFragment<FragmentInfoReviewDetailReply
 
             // 댓글 삭제 신고 바텀 시트
             ivInfoReviewCommentDot.setOnClickListener {
-                showInfoReviewCommentDeleteBottomSheet(infoReviewDetailReplyArgs.comment.itemId)
+                val commentUserName = infoReviewDetailReplyViewModel.infocommentdetail.value?.username ?: ""
+                showInfoReviewCommentDeleteBottomSheet(infoReviewDetailReplyArgs.comment.itemId, commentUserName)
             }
 
             // 댓글 상세
             val repository = CommentRepositoryImpl(RetrofitClient.commentApi)
-            val factory = ViewModelFactory { InfoReviewDetailReplyViewModel(repository) }
+            val factory = ViewModelFactory { InfoReviewDetailReplyViewModel(mainViewModel.tokenPair.value.first.toString(),repository) }
             infoReviewDetailReplyViewModel = ViewModelProvider(this@InfoReviewDetailReplyFragment, factory)[InfoReviewDetailReplyViewModel::class.java]
             infoReviewCommentDetailViewHolder = InfoReviewCommentDetailViewHolder(binding)
 
             // 댓글 상세 요청
+            val reviewId = infoReviewDetailReplyArgs.review
             val commentId = infoReviewDetailReplyArgs.comment.itemId
-            infoReviewDetailReplyViewModel.getInfoCommentDetail(commentId, commentId)
+            infoReviewDetailReplyViewModel.getInfoCommentDetail(reviewId, commentId)
             infoReviewDetailReplyViewModel.infocommentdetail.observe(viewLifecycleOwner) { comment ->
                 infoReviewCommentDetailViewHolder.bind(comment)
             }
@@ -83,7 +82,12 @@ class InfoReviewDetailReplyFragment : BaseFragment<FragmentInfoReviewDetailReply
             tvUploadCommentReply.setOnClickListener {
                 val reply = editUploadCommentReply.text.toString().trim()
                 if (reply.isNotEmpty()) {
-                    infoReviewDetailReplyViewModel.addReply(reply)
+                    val reviewId = infoReviewDetailReplyArgs.review
+                    val parentId = infoReviewDetailReplyArgs.comment.itemId
+
+                    infoReviewDetailReplyViewModel.postReply(reviewId, reply, parentId) {
+                        infoReviewDetailReplyViewModel.getInfoCommentDetail(reviewId, parentId)
+                    }
 
                     editUploadCommentReply.text?.clear()
 
@@ -100,25 +104,47 @@ class InfoReviewDetailReplyFragment : BaseFragment<FragmentInfoReviewDetailReply
                     if (item == null) {
                         return
                     }
-                    InfoReviewCommentDeleteBottomSheet(
-                        commentItemId = item.itemId,
-                        onDeleteConfirmed = { deleteItemId ->
-                            infoReviewDetailReplyViewModel.deleteReply(deleteItemId)
-                        }
-                    ).show(parentFragmentManager, "delete reply")
+                    if (item.isMine) {
+                        InfoReviewCommentDeleteBottomSheet(
+                            commentItemId = item.itemId,
+                            onDeleteConfirmed = { deleteItemId ->
+                                infoReviewDetailReplyViewModel.deleteReply(deleteItemId)
+                            }
+                        ).show(parentFragmentManager, "delete reply")
+                    } else {
+                        InfoReviewCommentReportBottomSheet(
+                            commentItemId = item.itemId,
+                            onReportConfirmed = { reportedId ->
+                                // 신고 후 처리
+                            }
+                        ).show(parentFragmentManager, "report reply")
+                    }
                 }
             })
             handleReplyUploadArea()
         }
     }
 
-    private fun showInfoReviewCommentDeleteBottomSheet(commentItemId: Int) {
-        InfoReviewCommentDeleteBottomSheet(
-            commentItemId = commentItemId,
-            onDeleteConfirmed = { deleteItemId ->
-                infoReviewDetailReplyViewModel.deleteReply(deleteItemId)
-            }
-        ).show(parentFragmentManager, "delete comment")
+    private fun showInfoReviewCommentDeleteBottomSheet(commentItemId: Int, commentUserName: String) {
+        val currentUserName = infoReviewDetailReplyViewModel.getUserNameFromToken()
+        if (commentUserName.equals(currentUserName, ignoreCase = true)) {
+            InfoReviewCommentDeleteBottomSheet(
+                commentItemId = commentItemId,
+                onDeleteConfirmed = { deleteItemId ->
+                    infoReviewDetailReplyViewModel.deleteReply(deleteItemId){
+                        findNavController().popBackStack()
+                    }
+                }
+            ).show(parentFragmentManager, "delete comment")
+        } else {
+            // 본인이 아니면 신고 BottomSheet
+            InfoReviewCommentReportBottomSheet(
+                commentItemId = commentItemId,
+                onReportConfirmed = { reportedId ->
+                    // 신고 후 처리
+                }
+            ).show(parentFragmentManager, "report comment")
+        }
     }
 
     // 대댓글 게시 입력창 위치 조정
